requirements.txt
langchain==0.1.8
langchain-cli==0.0.20
python-dotenv==1.0.1
google-cloud-discoveryengine==0.11.12
langchain-community==0.0.38
langchain_google_genai==1.0.3
pymongo==4.7.2
langchain-mongodb==0.1.5
python-multipart==0.0.9
pypdf==4.2.0
langfuse==2.33.0
pyjwt==2.3.0
cryptography==42.0.7